requirements.txt
# Core
python-dotenv
pydantic-settings

# LangChain
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-postgres
langchain-openai
langchain-google-genai

# Database
psycopg[binary]
sqlalchemy

# PDF Processing
pypdf

# Web Interface
chainlit

# Testing
pytest
pytest-asyncio

# Type checking (optional)
mypy